requirements.txt
sqlalchemy
langchain-chroma
pydantic
langchain-classic
openai
numpy
chromadb
langchain
langchain-huggingface
langchain-community
langchain-text-splitters
langchain-core
rank_bm25
tiktoken
sentence_transformers
langchain_openai
pytest
# Database and CSV processing dependencies
sqlmodel
psycopg2-binary
pandas
python-dotenv
unidecode

# Search and vector embedding dependencies
rapidfuzz
langchain-community
